requirements.txt
amqp==5.3.1
asgiref==3.10.0
bidict==0.23.1
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.7
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
idna==3.10
kombu==5.5.4
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-engineio==4.12.3
python-socketio==5.14.1
redis==6.4.0
requests==2.32.5
simple-websocket==1.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
vine==5.1.0
wcwidth==0.2.14
wsproto==1.2.0
